backend/requirements.txt
fastapi==0.115.11
strawberry-graphql==0.281.0
motor==3.7.1
pymongo==4.14.1
pydantic-settings==2.8.1
python-multipart==0.0.20
uvicorn==0.34.0